Modif suite a review et organisation code
